refactor: RemainingVotePower
